fix(data): balance and writing to file errors
fix: Balance will now correclty display no timer running when there is
no recorded data for the given day instead of panicing out of bounds.
Separated the file creation and the file reading and writing so that the
file is first read and stored in a buffer and then closed, and reopened
with write priviledges to avoid the bug that was leaving certain
characters undeleted after a successful rebalance.